+++ src/client/services/makeLayoutsStaticService.ts
